feat : Seperate project modal hooks
