+++ wise-office-client/src/components/modal/ProjectModal.tsx
@@ -1,19 +1,10 @@
-import { getMembers } from "@/services/members";
-import {
-    getProjectById,
-    postProject,
-    updateProject,
-} from "@/services/projects";
-import { Member } from "@/types/member";
 import { ProjectInfo } from "@/types/project";
-import { useEffect, useRef, useState } from "react";
+import { useRef, useState } from "react";
 import { useRouter } from "next/router";
-import { toastMessage } from "@/lib/common/toastMessage";
-import { useProjects } from "@/store/useProjects";
-import { CreateProject } from "@/types/createProject";
 import SelectProjectMembers from "../project/SelectProjectMembers";
 import ProjectNameWithPeriod from "../project/ProjectNameWithPeriod";
 import Tab from "../project/Tab";
+import { useProjectModal } from "@/hooks/useProjectModal";
 type ProjectModalProps = {
     mode: "create" | "update";
     projectId?: number; // update일 때만 필요
@@ -29,172 +20,42 @@ export default function ProjectModal({
     onClose,
     onCreated,
 }: ProjectModalProps) {
-    const [projectTitle, setProjectTitle] = useState("");
-    const [startDate, setStartDate] = useState("");
-    const [endDate, setEndDate] = useState("");
-    const [content, setContent] = useState("");
-    const [searchText, setSearchText] = useState<string>("");
-    const [selectedMembers, setSelectedMembers] = useState<Member[]>([]);
-    const [manager, setManager] = useState<Member | undefined>();
-    const [members, setMembers] = useState<Member[]>([]);
     const modalRef = useRef<HTMLDivElement>(null);
     const router = useRouter();
     const [tab, setTab] = useState("프로젝트 정보");
-    const handleProjectTitleChange = (value: string) => {
-        if (value.length > 100) {
-            toastMessage.error("프로젝트 제목은 100자까지 입력 가능합니다.");
-            return;
-        }
-        setProjectTitle(value);
-    };
-    const [errors, setErrors] = useState({
-        projectTitle: "",
-        startDate: "",
-        endDate: "",
-        content: "",
-        selectedMembers: "",
-        manager: "",
+
+    const {
+        projectTitle,
+        startDate,
+        endDate,
+        content,
+        searchText,
+        members,
+        selectedMembers,
+        manager,
+        errors,
+        setStartDate,
+        setEndDate,
+        setSearchText,
+        setSelectedMembers,
+        setManager,
+        handleSubmit,
+        handleContentChange,
+        handleProjectTitleChange,
+    } = useProjectModal({
+        mode,
+        projectId,
+        router,
+        onCreated,
+        onClose,
+        setProjectInfo,
     });
-    useEffect(() => {
-        const fetchMembers = async () => {
-            const members = await getMembers();
-            setMembers(members);
 
-            if (mode === "update" && projectId && router.isReady) {
-                const project_old = await getProjectById(projectId);
-
-                setProjectTitle(project_old.projectTitle);
-                setContent(project_old.detail);
-                setStartDate(String(project_old.start).slice(0, 7));
-                setEndDate(String(project_old.end).slice(0, 7));
-
-                setManager(
-                    members.find(
-                        (member) =>
-                            member.memberId ===
-                            project_old.managerName.memberId,
-                    ),
-                );
-
-                const selected = members.filter((member) =>
-                    project_old.attendant.some(
-                        (att) => att.memberId === member.memberId,
-                    ),
-                );
-
-                setSelectedMembers(selected);
-            }
-        };
-
-        if (!router.isReady) return;
-        fetchMembers();
-    }, [mode, projectId, router.isReady]);
-    const validateForm = () => {
-        let valid = true;
-        const newErrors = {
-            projectTitle: "",
-            startDate: "",
-            endDate: "",
-            content: "",
-            selectedMembers: "",
-            manager: "",
-        };
-
-        if (projectTitle.trim() === "") {
-            newErrors.projectTitle = "프로젝트명을 입력해주세요.";
-            valid = false;
-        } else if (projectTitle.length > 100) {
-            newErrors.projectTitle =
-                "프로젝트 제목은 100자까지 입력 가능합니다.";
-            valid = false;
-        }
-
-        if (startDate === "") {
-            newErrors.startDate = "시작 날짜를 선택해주세요.";
-            valid = false;
-        }
-
-        if (endDate === "") {
-            newErrors.endDate = "종료 날짜를 선택해주세요.";
-            valid = false;
-        }
-
-        if (startDate && endDate && startDate > endDate) {
-            newErrors.endDate = "시작일 이후로 선택해주세요.";
-            valid = false;
-        }
-
-        if (content.trim() === "") {
-            newErrors.content = "프로젝트 설명을 입력해주세요.";
-            valid = false;
-        } else if (content.length > 500) {
-            newErrors.content = "프로젝트 설명은 500자까지 입력 가능합니다.";
-            valid = false;
-        }
-
-        if (selectedMembers.length === 0) {
-            newErrors.selectedMembers = "프로젝트 참여 인원을 선택해주세요.";
-            valid = false;
-        }
-
-        if (!manager) {
-            newErrors.manager = "프로젝트 관리자를 선택해주세요.";
-            valid = false;
-        }
-
-        setErrors(newErrors);
-        return valid;
-    };
-    const handleSubmit = async () => {
-        if (!validateForm()) return;
-
-        const start = startDate + "-01";
-        const [year, month] = endDate.split("-").map(Number);
-        const lastDay = new Date(year, month, 0).getDate();
-        const end = `${endDate}-${lastDay.toString().padStart(2, "0")}`;
-
-        const projectData: CreateProject = {
-            projectTitle,
-            start,
-            end,
-            content,
-            projectManagerId: manager?.memberId,
-            attendants: selectedMembers.map((m) => m.memberId),
-        };
-
-        try {
-            if (mode === "create") {
-                const newProject = await postProject(projectData);
-                useProjects.getState().addProject(newProject);
-                if (onCreated) await onCreated();
-                toastMessage.success("프로젝트가 등록되었습니다.");
-            } else {
-                const updated = await updateProject(projectData, projectId!);
-                if (setProjectInfo) setProjectInfo(updated);
-                toastMessage.success("프로젝트가 수정되었습니다.");
-            }
-
-            onClose();
-        } catch (err) {
-            toastMessage.error(
-                mode === "create"
-                    ? "프로젝트 등록에 실패했습니다."
-                    : "프로젝트 수정에 실패했습니다.",
-            );
-        }
-    };
-    const handleContentChange = (value: string) => {
-        if (value.length > 500) {
-            toastMessage.error("프로젝트 설명은 500자까지 입력 가능합니다.");
-            return;
-        }
-        setContent(value);
-    };
     return (
         <div className="Overlay fixed inset-0 bg-opacity-40 flex justify-center items-center z-50 p-6">
             <div
                 ref={modalRef}
-                className="bg-white rounded-xl shadow-xl p-8 w-full max-w-[64rem] max-h-[76vh] flex flex-col relative"
+                className="bg-white rounded-xl shadow-xl p-8 w-full max-w-[64rem] h-[76vh] flex flex-col relative overflow-auto"
             >
                 {/* 닫기 버튼 */}
                 <button
@@ -221,8 +82,7 @@ export default function ProjectModal({
                     </div>
                 </div>
                 {/* 좌우 영역: flex-grow 해서 남은 높이 전부 차지 */}
-                <div className="flex flex-col md:flex-row gap-8 flex-grow overflow-hidden">
-                    {/* 왼쪽 영역 */}
+                {tab === "프로젝트 정보" ? (
                     <ProjectNameWithPeriod
                         projectTitle={projectTitle}
                         startDate={startDate}
@@ -234,8 +94,7 @@ export default function ProjectModal({
                         setContent={handleContentChange}
                         errors={errors}
                     />
-
-                    {/* 오른쪽 영역 */}
+                ) : (
                     <SelectProjectMembers
                         members={members}
                         searchText={searchText}
@@ -246,7 +105,8 @@ export default function ProjectModal({
                         setManager={setManager}
                         errors={errors}
                     />
-                </div>
+                )}
+                {/* </div> */}
 
                 {/* 생성 완료 버튼 */}
                 <button
